Modularize NewMessageStackNavigator
